modification func calculate

+++ CountOnMe/Extension/Float+Extension.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-extension Float {
+extension Double {
     var removeZeroAfterComma: String {
        return self.truncatingRemainder(dividingBy: 1) == 0 ? String(format: "%.0f", self) : String(self)
     }
